Move to Spring 3.2.0.RELEASE and fix relevant imports
