+++ pom.xml
@@ -18,9 +18,8 @@
 		<mockito.version>1.9.5</mockito.version>
 		<servlet.version>3.0.1</servlet.version>
 		<slf4j.version>1.6.6</slf4j.version>
-		<spring.version>3.2.0.BUILD-SNAPSHOT</spring.version>
+		<spring.version>3.2.0.RELEASE</spring.version>
 		<spring-hateoas.version>0.3.0.RELEASE</spring-hateoas.version>
-
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 	</properties>
 
@@ -116,16 +115,4 @@
 			</plugin>
 		</plugins>
 	</build>
-
-	<repositories>
-		<repository>
-			<id>spring-snapshot</id>
-			<name>Spring Snapshot Repository</name>
-			<url>http://repo.springsource.org/snapshot</url>
-			<snapshots>
-				<enabled>true</enabled>
-			</snapshots>
-		</repository>
-	</repositories>
-
 </project>
